#     Faster searches for "sift -w WORD" word boundary searches.