--- a/matching.go
+++ b/matching.go
@@ -21,6 +21,7 @@ import (
 	"io"
 	"os"
 	"regexp"
+	"regexp/syntax"
 	"sort"
 )
 
@@ -173,6 +174,24 @@ func processReader(reader io.Reader, matchRegexes []*regexp.Regexp, data []byte,
 					if (!options.Multiline && m.lineEnd > prevMatch.lineEnd) ||
 						(options.Multiline && m.start >= prevMatch.end) {
 						validMatch = true
+					}
+				}
+				// work around the regexp engine when searching for a word flanked by word boundaries: \bWORD\b.
+				// When -w is set, omit the leading word boundary \b in the hope that WORD starts with a string literal,
+				// which can be quickly searched for before entering the slower regex engine.
+				// Enforce the leading word boundary requirement here. \b == \A\w or \W\w or \w\W or \w\z
+				// \w\z is out as a possibility because the pattern has a trailing \b
+				if options.WordRegexp && validMatch {
+					if syntax.IsWordChar(rune(newMatches[i].line[newMatches[i].start-newMatches[i].lineStart])) {
+						// \A\w or \W\w
+						if !(newMatches[i].start == newMatches[i].lineStart || !syntax.IsWordChar(rune(newMatches[i].line[newMatches[i].start-1-newMatches[i].lineStart]))) {
+							validMatch = false
+						}
+					} else {
+						// \w\W
+						if !(newMatches[i].start > newMatches[i].lineStart && syntax.IsWordChar(rune(newMatches[i].line[newMatches[i].start-1-newMatches[i].lineStart]))) {
+							validMatch = false
+						}
 					}
 				}
 				if validMatch {