Equivalent to matching the regex \bWORD\b.

    Shift the order in which characters of the haystack are compared to the needle. Instead of starting the search on a word boundary (\b), start with WORD\b. When WORD has a string literal prefix (a common use case), searching for it is more efficient because it occurs less frequently (in most texts) and is faster because it bypasses the regex engine [1]. As a second step, enforce the leading word boundary condition over a much smaller haystack, without a regex.

    [1] https://github.com/golang/go/blob/a064a4f29a97a4fc7398d1ac9d7c53c5ba0bc646/src/regexp/backtrack.go#L341

    TL;DR it's faster to search for literals first.

    Observed 16x speedup for the following pattern and directory:
    sift -q "\bWaitForConnect\b" $m/mist-ap  2.62s user 0.08s system 555% cpu 0.486 total
    sift -w -q "WaitForConnect" $m/mist-ap  0.08s user 0.10s system 570% cpu 0.032 total

    ** HOWEVER, there is a bug in sift that fails to find any matches for patterns that start with "\b\W". ***